% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suportme.R
\name{supportme}
\alias{supportme}
\title{Gives Support When Needed}
\usage{
supportme(type = c("love", "encouragement", "humor"))
}
\arguments{
\item{type}{One of three support types "love", "encouragement", or "humor"}
}
\description{
Use this when you encounter an error in R. It may help
}
